+++ Basalt/Engine.cs
@@ -176,6 +176,7 @@ namespace Basalt
 			Instance.Logger?.LogDebug($"Creating entity {entity.Id}...");
 			Instance.EntityManager.AddEntity(entity);
 			Instance.GetEngineComponent<IPhysicsEngine>()?.AddEntityToSimulation(entity);
+			entity.created = true;
 			entity.CallStart();
 		}